+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------------
 
   * new Backend WPTrackserver directly accessing its mysql database
+  * new: Backend.clone()
 
 1.3.2 release 2018-08-19
 ------------------------